Boneka

+++ src/main/java/com/TokoBoneka/TokoBoneka/impl/ProdukImpl.java
@@ -56,7 +56,6 @@ public class ProdukImpl implements ProdukService {
         produk.setDeskripsi(produkDTO.getDeskripsi());
         produk.setHarga(produkDTO.getHarga());
         produk.setStok(produkDTO.getStok());
-        produk.setImageUrl(produkDTO.getImageUrl());
 
         Produk savedProduk = produkRepository.save(produk);
 
@@ -67,7 +66,6 @@ public class ProdukImpl implements ProdukService {
         result.setDeskripsi(savedProduk.getDeskripsi());
         result.setHarga(savedProduk.getHarga());
         result.setStok(savedProduk.getStok());
-        result.setImageUrl(savedProduk.getImageUrl());
 
         return result;
     }
@@ -85,7 +83,6 @@ public class ProdukImpl implements ProdukService {
         existingProduk.setDeskripsi(produkDTO.getDeskripsi());
         existingProduk.setHarga(produkDTO.getHarga());
         existingProduk.setStok(produkDTO.getStok());
-        existingProduk.setImageUrl(produkDTO.getImageUrl());
 
         Produk updatedProduk = produkRepository.save(existingProduk);
 
@@ -96,13 +93,16 @@ public class ProdukImpl implements ProdukService {
         result.setDeskripsi(updatedProduk.getDeskripsi());
         result.setHarga(updatedProduk.getHarga());
         result.setStok(updatedProduk.getStok());
-        result.setImageUrl(updatedProduk.getImageUrl());
 
         return result;
     }
 
     @Override
     public void deleteProduk(Long id) throws IOException {
+        if (!produkRepository.existsById(id)) {
+            throw new IllegalArgumentException("Produk dengan ID " + id + " tidak ditemukan.");
+        }
         produkRepository.deleteById(id);
     }
+
 }
